feat(connections): fold default messaging channel into Channels tab
The Settings → Messaging Channels page had two halves:
- 'Default messaging channel' picker (the part the user actually used)
- Channel connection list (duplicated by the Channels tab in /skills)

Move the default-channel picker into the Channels tab under /skills as
a footer section, and delete the Settings page outright:

- DefaultChannel picker now lives at the bottom of the Channels card,
  reusing the same channelConnectionsApi.updatePreferences flow.
- Remove MessagingPanel.tsx + its two test files + the /settings/messaging
  route + the Developer Options tile that pointed at it.
- Redirect the Rewards community CTA from /settings/messaging to /skills
  (closest equivalent — same affordance, fewer hops).
- Leave the /settings/messaging path-matching in BottomTabBar +
  useSettingsNavigation as harmless dead branches (URL no longer
  resolves).
