+++ app/src/components/rewards/RewardsCommunityTab.tsx
@@ -110,7 +110,7 @@ export default function RewardsCommunityTab({
           </div>
           <div className="flex flex-col gap-2 sm:flex-row">
             <button
-              onClick={() => navigate('/settings/messaging')}
+              onClick={() => navigate('/skills')}
               className="inline-flex items-center justify-center gap-2 rounded-xl bg-white dark:bg-neutral-900 px-4 py-3 text-sm font-semibold text-primary-700 dark:text-primary-300 shadow-lg transition-transform active:scale-[0.98]">
               <svg
                 className="w-4 h-4"
